requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
click==8.1.4
Django==4.2.3
django-debug-toolbar==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rest-framework-simplejwt==0.0.2
rpds-py==0.8.10
sqlparse==0.4.4
uritemplate==4.1.1
